Stop delegating to roboleaf mode if command has unsupported flags
Sometimes users specify flags which are not supported by roboleaf mode
and don't suppose to use roboleaf mode, even though the test is
supported by roboleaf mode. For example, the user wants to get a
coverage report and specifies `--experimental-coverage`, but the
feature hasn't been supported by roboleaf mode.

Note that this change creates a draft list of unsupported flags, but
we'll keep cleaning up the list in following CLs.

Bug: 292200021
Test: m atest && atest-dev HelloWorldHostTest -b
Test: m atest && atest-dev HelloWorldHostTest
Test: m atest && atest-dev HelloWorldHostTest -v
Change-Id: Ic07fdc8ebda3cba8dfc6152d9049156aa3bcdfc7