Refine logic of searching modules from path
Previously Atest determined modules to test from a given path by
searching which module located in that path, which led incomplete
searching result when the given path is also defined in the `srcs` of
another testable modules.

This change fixes the bug by checking whether the given path defined in
the `srcs` of the module; if not, checkes whether the given path belongs
to a module's path. Also, this change speeds up returning modules to
test by narrowing down the searching scope to testable modules only.

Bug: 298089212
Test: atest DreamingToLockscreenTransitionViewModelTest # no selection
Test: atest-dev DreamingToLockscreenTransitionViewModelTest # selections
Change-Id: Ifa85b610072ab8bb159fd773335af8efee0392b2