Add Permission bits to file fingerprint.
Adding permission/mode bits to the json being sent back from the device and
computed on the host.
However, we are specifically clearing the bits before doing comparison.
We avoid the problem of the current dozen users having an old device
fingerprint tool that doesn't emit the bits and compare with the host
json which does have it.
I increased the version number of the tools and will increase again next
release so we can use that to figure out if we should enable the check.

Test: atest
Test: CI
Test: adevice status
  # See the permission bits
Test: adb shell /system/bin/adevice_fingerprint -p system/fonts | jq
Test: adb shell /system/bin/adevice_fingerprint --version
  # Test with old version of adevice_fingerprint on device.
Test: m adevice && git checkout main &&  m adevice_fingerprint && adb push $ANDROID_PRODUCT_OUT/system/bin/adevice_fingerprint system/bin/adevice_fingerprint && adevice status

Change-Id: I391a114060f427beac6a44e900b48c94cce8a836