Remove Unused Parameters in extra_args_to_tf_args
This change removes the 'mod_info' and 'test_infos' parameters from the
extra_args_to_tf_args function, as they are not used within the
function. These parameters were originally included in the function
signature but were not utilized in any part of the function's logic.
Removing them simplifies the function and improves code clarity.

Bug: 301202610
Bug: 301187663
Test: m atest && atest-dev --host atest_unittests
Change-Id: I40fafc19e68604b0eace4b42dad388d415eee17a