Fix the Java test filter validation in test info cache finder
ATest validates the Java test filters in cache by comparing the test
class filters with the source file paths. But the class package won't
be always aligned with its source file path.

Also, there is a logic defect that the test info is valid as long as
one test filter is valid.

Bug: 172260100
Bug: 301328077
Test: m atest && atest-dev HelloWorldTest
Test: atest-dev hello_world_test:HelloWorldTest --roboleaf-mode=off
Test: atest-dev CarDevicePolicyManagerUnitTest
Change-Id: I3551d268c0100dac078ee09c234977c241595b75