Move lots of output to log file.
All the "adb" commands now go to the logfile instead of the console.
If we are pushing a lot of files, it might sit there for a while :(

Changed logger to use tracing_logger.

Tracing_logger is a superset of logger.
It allows us to log to files easier and also has macros and derives for
instrumenting methods (so we can later remove our own).

Always writes log files -> $ANDROID_BUILD_TOP/out/adevice.log
Continues gracefully if logs can not be written.

Test: atest
Test: # modify file; adevice update

Change-Id: I09e4df2185725a00fdc68e0959af2dc88d109090