Reduce complexity of Test Mapping related code in `main`
Atest executes Test Mapping and regular tests differently. The code to
check for which type of execution is repeated and spread throughout
the main function which makes code hard to understand and maintain.

This change introduces an abstraction that splits the different
execution paths into subclasses and simplifies the call sites.

Test: atest-dev hello_world_test
Test: atest-dev --test-mapping tools/tradefederation/core
Test: atest-dev --test-mapping tools/tradefederation/core --device-only
Test: atest-dev --test-mapping tools/tradefederation/core --host
Change-Id: I8901ec651431ba15186f75e8128334d669783e04