Make `_run_limited_output` suitable for general use
The `_run_limited_output` function assumes it is used for the build
and cannot be used for streaming lines of other commands. This is
because it writes to build-specific log files in the output directory.

This change extracts a general version of that utility for use by
other non-build Atest uses.

Bug: 314339352
Test: atest-dev atest_unittests
Test: manually break the build
Change-Id: I5169edff6448e65bec907d2ddd163d274d1d0d6b