Integration test runs with clean env vars
Previously the integration test is running with restored env vars + existing env vars in the OS. The atest command would fail to run in the test lab because the host machines has env vars defined such as $TF_GLOBAL_CONFIG that would affect the atest execution. This CL cleans up the env vars so that the atest commands executes in a clean env.

Test: test: python tools/asuite/atest/integration_tests/atest_ci_tests.py --artifact_pack_path /tmp/a/atest_integration_tests.tar -t -s <serial>
Test: test: atest atest_ci_tests --  --test-arg com.android.atest.IntegrationTestArtifactPathOptionProxy:artifact-pack-path:/tmp/a/atest_integration_tests.tar
Bug: 246486566
Change-Id: Ic1e4557e6382beb2eae63b33ca4b88d0024746fe