Increment file's modified time instead of sleeping in Snapshot test
Previously, the test relied on `time.sleep` to ensure that the file's modified timestamp was different after the snapshot was taken. However, this approach is not deterministic and can lead to flaky tests.

By incrementing the file's modified time directly using `os.utime`, we ensure that the test consistently detects the file modification without introducing unnecessary delays.

Test: atest atest-unittests
Change-Id: I6a7423047928f5bd295613556696eb68c91b3981