Apply the new inject-build-key Tradefed feature
Previously we created the IntegrationTestArtifactPathOptionProxy in order to pass artifact path to the python host test, and filed a feature request to tradefed to add native support this. Now the new TF feature has been merged, this CL removes our IntegrationTestArtifactPathOptionProxy solution and apply the new TF feature 'inject-build-key' to inject the artifact path.

Test: go/run-atest-integration-tests
Bug: 316223008
Fix: 319483899
Change-Id: I7b82ae0e80bfb285dd9e0deea56823549c32ae56