Compress snapshot files to reduce artifact size by half
Individually compress the files in snapshot storage to reduce artifact size. Note that we are not compressing the files while creating the blob store because the blob store creation is harder to be parallelized. We are also not compressing the directory directly because tarfile module doesn't support parallel mode. The approach in this CL that compress individual files after creating the blob store and then tar the entire snapshot storage is fastest and results in around 50% artifact size reduction while only adding around 30 seconds to the build and test steps.

Test: atest atest_integration_tests
Change-Id: I546194010c61c338bd8055d64b588ebb292d5784