Build atest-dev inside the python script
Building atest-dev is a necessary part of the integration test because we are testing the binary atest-dev instead of the atest source code. Previously the binary is built in the shell script. But now we support directly running the test via python or atest, we should move the build call to the python script so that the integration test is complete.

Test: atest atest_integration_tests
Bug: 318747715
Change-Id: I5bbd530768344eec5a65075b9fd75d59571a663f