Dedicate the integration shell script to build server invocations
We have 3 equivalent ways to invoke the integration test and it's confusing for the users to understand the difference and choose. Also the shell script has logic to determine the repo root that requires either the android build top be set or executed in the repo root path which a new user is usually unaware of. This CL renamed the shell script and added description so that the shell is script is dedicated for build server use. Users are redirected to either call the test via python or atest by the description.

We keep the old script for legacy build command support until the build server side command is updated.

Test: build_atest_integration_tests.sh
Bug: 318747715
Change-Id: I276c24d54249299eae7f2a459085c1cc089f26ce