Initial change to extract asuite common integration test logic from atest
Note that it's intentional that the new module is named to split_bulid_test_script while the class inside it is still AtestIntegrationTest so that we don't have to rename the module again later when we convert the class to script style. The new module still contain many atest specific logic and they will be moved out in the coming CLs.

Test: atest atest_integration_tests
Bug: 319930315
Change-Id: I508b1f592956c2ed928a4b94bc757c445291bf67