Extract individual integration test classes to their own modules
As our test class is getting longer and more class is bening planned to be added, it's time to split the integration tests class into individual test classes. This improves readability and allows running individual tests with the atest command. For example, deviceless integration tests can be run with the --host param.

Test: atest atest_integration_tests
Test: atest atest_command_verification_tests --host
Bug: 319930315

Change-Id: I4b4c6311ed0ab2ada324a82dd9360f29f33d2925