Improve error message when atest command failed
Previously we use CompletedProcess's built in method check_returncode() to check atest command's return code, but that method only prints the return code is non-zero without telling why the atest command failed. The missing information is always in stdout. After this CL, the stdout will be included in the exception's error message when return code is non zero.

Test: atest atest_integration_tests
Bug: 321796949
Change-Id: I0c5e69a30ecc570dfec1ab9c87f76a488f193bef