Script to filter all_modules from build to just tests.
The build target "all_teams" is writing a proto file that has "team"
information for all modules.
This script filters it and writes a new proto that just has test
modules. "test" modules isn't well defined by soong, so we use a
heuristic and read module-info.json. Otherwise we need to check
the rule kind in song for "endswith(_test)" which seems wrong too.

As a bonus, this converts to the indentical proto all_test_specs in the
output.

Test: m filter_teams && out/host/linux-x86/bin/filter_teams  --filter_teams
Change-Id: Ifffe72366dfecafc44d79386a0c0b507c5dce0ac