Continue running linter on all files before exiting on lint issues
Previouly, when one file has lint issue, the preupload hook script would print the issue and stop running. This is inefficient when people are uploading multiple files with lint issues. This CL updates the script to make the script continue running linter on all files even when a lint issue is found.

Test: repo upload .
Change-Id: Idedc78762efe5bb93adcbdff3bb219697e004fa5