Print atest result dir on each run
The atest result dir contains test results and logs and is different on every run (created as a temp dir). Therefore printing it to console is the only way to lead either people or tests to access the results and logs. Previously it is not printed and we need to extract it from the log html path, but the log html isn't generated in every scenario and parsing the path also makes tests brittle. This CL prints the results and logs directory once they are created, so that test logging can be added to anywhere in the code base and no test logs will go missing.

Test: atest atest_integration_tests
Fix: 322860039
Change-Id: I76d53f3a8fc9794719ccf4a3f4d657ee881e9be2