Minor refactor on atest command verification tests
Combine the test name and test arg param in the _verify_atest_internal_command method. Previously we need to split test name and test args as separate params because we need the test name to get the runner commands from the test_commands dict. Now it's no longer needed as we are using --dry-run outputs.

Test: python atest_command_verification_tests
Change-Id: Ie02da8afbcd3c4d95bf2e6771f1c92868bbfd375