Executes command verification tests in parallel
Command verification tests are no disk operation tests and can be run in parallel. This CL created an annotation to mark tests for parallel run and make the command verification tests all running in parallel.

The test speed is now 30X times faster and complete in just a few seconds.

Test: run the command verification tests locally and in forrest
Fix: 323005540
Change-Id: Ia4421ddd1c14ef09ff498eec37759358638118cc