Support parsing multi-line atest log entries
Previously since atest logs only contains single line entries and the integration test parse the logs with the same assumption. However as we are adding more logs to atest, multi-line logs are possible and the original impl would thrown an error if encountered one. This Cl removed the assumption and makes the integration able to parse multi-line logs.

Test: add a multi-line atest log, and run the integration tests
Change-Id: I2266c62dc47d6b9d2d3f63d5282fb30cb700f7b0