Verify atest commands to be equivalent instead of subset
Previously the command verification tests are checking whether the expected command args are a subset of the actual command args. This is not effective because the actual command containing unwanted arguments will still pass the test. For example, the runner command for deviceless tests contains extra arguments like '-n'. If atest mistakenly treats a device test into a deviceless test and added '-n' to the runner command, the command verification will not catch it. This CL added back the command normalizing logic from the original atest -y verification to sanitize the runner commands and perform set equal instead of subset check.

Test: Locally modified the test case to verify that when a device test is treated as deviceless test the command verification would fail.
Fix: 324476085
Change-Id: I8801c927832ab04bcbf36284af627ac237f074aa