Fix the command verification integration tests
When the command verification tests runs in parallel it expects no build action is needed. This is true locally because we invoke building atest before the test so that the module-info is built already. But the lab environment is setup differently and building atest doesn't build the module info. When the command verification runs in parallel, all threads tries to build the module-info simutaniously, resulting in lock timeout from soong. This CL adds a setupClass method to the tests so that a atest command is executed before the parallel run and build the module-info along with any other dependencies.

Fix: 327712963
Test: forrest
Change-Id: I43be83d684ffd04e4b26c08785fc16b11fae4324