Add unit tests for module and class finders and add fix module not found exception.
Add unit tests that rely less on mocks for module finder methods.
Also add an additional check on the returned module info to avoid a not found exception in case an bad module name is passed.
There is still a bug for finding tests by Module:Class because it's not using the correct config for the second run of the test. Details in bug.

Bug: 328148643
Test: m atest && atest-dev HelloWorldTests
Test: atest-dev HelloWorldTests:HelloWorldTest
Test: atest atest_integration_tests
Change-Id: Ic29403dd23f555d456e01e080d97f895c33cc4d8