Convert the 'Wrong java version' error into a warning
Since atest is run by developers from within the Android tree,
usually the platform Java toolchain is used. If that is not the
case, it is enough to inform that the version is not one of the
known ones without stopping the test run. This allows usage of
newer Java versions without changes to the script.

Test: TH
Change-Id: I782c6e4325e3766a643f48527f353489953aac1b