Fix up the request-upload-result feature
The documentation for the request-upload-result states that the option only need to be called once and will apply to subsequent runs, but the current implementation is that it only applies to the current run. This is due to b/275113186 and the function to keep the flag state for the subsequent runs is temporarily disabled in aosp/2537472. Now since the bug has been fixed, this CL aims to restore the changes made in aosp/2537472.

Test: run atest with the log uploading enabled and disabled.
Change-Id: Idbb8e962d2785b402d9d067371979c41e10d6bc1