Replace the bazel mode tests with non bazel mode in atest integration tests
The bazel mode integration test was broken recently, making the test fail. Since bazel mode has its own integration tests and it's also going to be deprecated in the long term, it's a of lower priority to fix the support of testing it in the atest integration tests. This CL replaces the tests with non-bazel mode so that our tests can be green again to monitor other breakages.

Test: atest atest_integration_tests
Fix: 330599501
Change-Id: I48b3f823bf9306a34eef808d9e4dad522870a439