Load vendor config from json if available in the environment
This is part of the work to eventually replace the dangling import of constants_google entirely. We will start with migrating the constants to json and source code. In the future, this json config will also be removed, and the configuration will be received either from environment variable or command flags.

Test: run atest with and without the environment variable.
Bug: 333600735
Bug: 323532099
Change-Id: Ie0f917a297d90dbf7a0fd9d23a1f5b23bc595d0d