Pass the test_filter from the user args to the Atest TF runner
The test_filter passed by the user doesn't make its way into the test_info objects for tests. This causes the test_filter to be ignored.
Bypass the Test Infos and add the filter to `--atest-include-filter` in the atest_tf_runner so that it can be added to the corresponding test runners by TF.

Bug: 315033800
Test: atest-dev python_host_test_example --test-filter .*pass.*
Test: m atest && atest-dev atest_unittests
Test: atest-dev HelloWorldTests --test-filter .*Welt.*
Test: atest-dev HelloWorldTests:HelloWorldTest#testHalloWelt
Test: atest-dev HelloWorldTests:HelloWorldTest#testHelloWorld --test-filter .*Welt.*
Test: atest-dev atest-dev CtsDeqpTestCases --test-filter dEQP-VK.api.smoke#*
Change-Id: I85c8f8e157193cdf2773c71972f7214830eca1fb