Detect USB speed for one device in an atest execution and print a warning
This uses the new adb command that allows us to detect the USB speed of the user's device.
We detect the USB cable speed, compare to the max speed the device is capable of, and if the USB cable is too slow we print a warning to the users and let them know they should use a faster cable. This can save users a lot of time when installing apps and flashing devices. You can read more in ADB team's design doc at go/android_usb_speed.
This only applies to local desktop users. We don't have enough info to detect for cloudtop users.

Bug: 341387238
Test: m atest && atest_unittests

Change-Id: I3de7862b83f916abc6359b2be042d4f43a51bc05