Fix module finders to not be greedy when module src matches test's file path.
Module finders try to find as many modules as possible based on the test's file path.
This causes workflow interruptions when several modules are found and users are presented with several module options to choose from.
If we already found a matching module that has the tests's SRC in its module-info, return that module immediately.

Bug: 315759481
Test: m atest && atest-dev atest_unittests
Test: atest-dev WindowInsetsAnimationImeTests -c
Change-Id: I8f5e807cf358cf4f57eac97831736235d184732a