Fix atest crashing when tf output file is not set
Currently atest process would crash when running an instrumentation test with -v and the instrumentation crashed. This CL fix the crash, but doesn't fix the warning printing feature when running with -v. The real fix to the feature is currently blocked and tracked by b/362606617. We need this fix to merge first so that the process will at least not crash.

Test: atest device_android_test_non_starting -v
Fix: 362612193
Change-Id: Id370359e87a9118617a764a3a736753dc8570547