Fix bad unit tests that overwrite global module constants
In one of the result reporter unit test case, the global module constant was modified to atest_configs.GLOBAL_ARGS without using mock or restoring the state. There are 100+ other test cases were written assuming atest_configs.GLOBAL_ARGS is defined as a mock object. The unittest was passing only because the bad test case was executed first, but when occasionally the test was not executed in the same order, any of those test cases executed before the bad test case will fail due to the constant's original value is None.

Fix: 365264561
Test: atest atest_unittests
Change-Id: I2bb84bdbc4a5782ac0e126eb4d227a8b4fa219e6