Make device update an regular arg instead of a step
The steps defined in the main class should be executed when True, but the device update step has additional conditions to satisfy before it should be executed. It's not a step but more of a feature like the start acloud/avd before tests. This CL removes device update from the Steps class to improve readibility.

Test: atest_unittests
Bug: 367485992
Change-Id: I0c56126e952e8fe237dd300354805fcea7af06d9