Make _configure_update_method explicitly return dependencies
Previously, _configure_update_method stores the dependencies in the device_update_method. Not only is it confusing that a 'method' has a 'dependencies' function, but also in the build step the dependencies gets silently injected into the build target list. It was not intruitive that the configure_update_method should be called before the build step. This CL makes the _configure_update_method method to explicitly return the build dependencies.

Test: atest_unittest
Bug: 367485992
Change-Id: I0cca3bd6c764b2aea8eace44d28c0f64f2f3a2a6