Fix extra erase of 2 output lines when rolling print is still empty
The reset function erase (len(last_lines) + 2) number of lines, where the extra 2 lines is for the starting and ending line such as '========== Rolling subprocess output =========='. When last_lines is still empty, these lines haven't been printed to the screen, and thus should not be erased. This CL adds a check for the emptiness of the rolling print buffer to avoid extra erasing.

Test: atest unittests
Change-Id: I79f72bb4cdf580e052ea88d322d6b400a1c915b4