Do not print rolling output feature message if not using it
Currently the TF runner checks whether the rolling output feature is enabled first. If it returns True, the feature message is printed. However, the feature being enabled does not mean it will be used as the runner has 2 other checks following the first check. This results in when user runs a test with --verbose, there's no rolling output but the feature message was printed. This CL changes the check order to fix this.

Test: make atest, atest-dev
Change-Id: I97e7d06c66f2228ce2ef7671e5d1ea6344c25cf6