atest --experimental-coverage: merge profiles before generating report
llvm-cov can only handle a single --instr-profile parameter as input. In
case multiple inputs are passed, only one of those is processed.

atest assumes it can pass multiple of them, and as a result generates a
report with partial coverage data.

To fix this, we can merge all input profdata files into a single one,
and pass only that single file to the llvm-cov command.

Bug: 423814386
Test: atest-dev test-binary-1 test-binary-2 --experimental-coverage
Change-Id: I5e3c378f1d00eee577ed7dcc855685acc7d591ef