Tolerate file find command failure
On cog, atest uses find command to find some config files. But the
command may return -1 when the find command failed to access some
special directories, even if it was able to successfully find the target
file. Such failure stops the entire atest main logic as it throws error.
This CL fixes it.

Bug: 432839394
Change-Id: I4d777b622d8454630cdeb4eab3df92f2b276cc69