[Refactor] Remove redundant import in atest_main.py
Removed the direct import of  from
and updated the type hint in  to use
.

This change is purely a style improvement that makes the code more
consistent. The  module was already imported, so importing
 directly was redundant. Using  everywhere
improves readability and consistency.

This change does not alter the program's logic or functionality as it
only affects a type hint and an unused import.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: Ifad0059b208e46a9846daaff1eee509c92b2643f