[Refactor] Use f-string for AttributeError in get_extra_args
This change updates the string formatting for the AttributeError in the  function to use an f-string.

Using an f-string is a modern Python feature that is more readable and generally preferred over the older '%' formatting. This change improves code clarity and adheres to modern Python best practices.

This is a purely stylistic change and does not alter the logic of the program. The exception message will be formatted identically to the previous implementation, ensuring that no behavior is changed.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: I8008038b4fcb31d35169a828e7e44ec37937167b