[Refactor] Simplify boolean to integer conversion in _has_valid_test_mapping_args.
The original code used an if/else block to pass either 1 or 0 to the
`metrics.LocalDetectEvent` based on the boolean value of `is_test_mapping`.

This change refactors the code to directly convert the boolean `is_test_mapping`
to an integer using `int()`. This is a more concise and Pythonic way to
achieve the same result, as `int(True)` evaluates to 1 and `int(False)`
evaluates to 0.

This change is purely a stylistic improvement for readability and conciseness.
It does not alter the existing logic or behavior of the function.

Test: atest atest_unittests
OFF_PEAK_PRESUBMIT: true

Change-Id: Ie06d7d3660703eaaad852a74da34206aad510107